requirements.txt
numpy
opencv-python
pydub
scipy
Pillow
moviepy
ttkbootstrap
moderngl